Split logger-slf4j into an isolated jar
