+++ logger-slf4j/src/main/java/com/palantir/logsafe/logger/slf4j/Slf4jSafeLoggerBridge.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.palantir.logsafe.logger;
+package com.palantir.logsafe.logger.slf4j;
 
 import com.google.errorprone.annotations.CompileTimeConstant;
 import com.palantir.logsafe.Arg;
