+++ settings.gradle
@@ -5,5 +5,6 @@ include 'safe-logging-refactorings'
 include 'preconditions'
 include 'preconditions-assertj'
 include 'logger-generator'
+include 'logger-slf4j'
 include 'logger-spi'
 include 'logger'
